+++ src/time_split/support/__init__.py
@@ -13,6 +13,7 @@ dependencies if you need to use the ``support`` module.
 
 from .._backend import DatetimeIndexSplitter, expand_limits, process_available
 from .._frontend import default_metrics_formatter, fold_weight, to_string
+from ._create_explorer_link import create_explorer_link
 
 __all__ = [
     "DatetimeIndexSplitter",
@@ -21,4 +22,5 @@ __all__ = [
     "default_metrics_formatter",
     "fold_weight",
     "to_string",
+    "create_explorer_link",
 ]